backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
yt-dlp>=2023.11.16
supabase==2.0.0
# Phase 1: Karaoke processing dependencies
# Note: Using Demucs instead of Spleeter (better maintained, no dependency conflicts)
demucs>=4.0.0
openai-whisper>=20231117
pysubs2>=1.6.1

